pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pytorch-template-code"
version = "0.1.0"
description = "A clean generic PyTorch template for computer vision tasks."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "numpy>=1.26.4",
    "torchvision",
    "torchaudio",
    "tqdm",
    "pandas",
    "pyyaml",
    "wandb",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
